pr-471 (#32)
Import from: https://github.com/apache/mesos/pull/471

When egress rate limiting is applied to the existing container without it, the HTB qdisc is created with default class ID configured to 0 (default). Because of that egress traffic from the container misses the rate limiting class 1:1. This fixes it by setting default class ID to 1.

Co-authored-by: Ilya Pronin <ipronin@twitter.com>
Co-authored-by: Franck Cuny <fcuny@twitter.com>
Reviewed-on: https://gitea:3000/AVENTER/clusterd/pulls/32