Fixed sorting in MountInfoTable::read. (#36)
Import from: https://github.com/apache/mesos/pull/468

Current implementation of MountInfoTable sorting crashes when some parts of the mount tree are not visible in /proc/{pid}/mountinfo, which can happen, for example, in a chrooted environment. The sorter assumes that the / mount is always present and uses it as the starting node for the DFS traversal.

Additionally, this sorting algorithm loses mounts that are parents of themselves from the sorted MountInfoTable. The bug is not caught by FsTest.MountInfoTableReadSortedParentOfSelf test, so it needs to be updated as well.

The new sorting implementation supports multiple starting nodes and works in situations where the visible mount tree is disconnected. It also changes the resulting order to BFS, but that shouldn't matter because the invariant is that all parent entries appear before their child entries.

Co-authored-by: Ilya Pronin <ipronin@twitter.com>
Reviewed-on: https://gitea:3000/AVENTER/clusterd/pulls/36