pr-475 (#38)
Import form: https://github.com/apache/mesos/pull/475

If quotas are set for the XFS disk isolator they will only be
enforced if the limit enforcement feature is enabled, and it
is not always enabled by default. This change will auto-enable
quota enforcement on the XFS disk isolator, to get the expected
quota enforcement behavior.

Co-authored-by: Ilya Pronin <ipronin@twitter.com>
Co-authored-by: Franck Cuny <fcuny@twitter.com>
Reviewed-on: https://gitea:3000/AVENTER/clusterd/pulls/38